requirements.txt
pandas
apache-airflow-providers-mongo